Bits & Bobs
Very glad I did do a QuickSave commit. In my attempt to have another go at the A* script, unity froze on start and had to be shutdown by task manager on 2 separate occasions. Lost progress.

- Pathfinder no longer slips diagonally through gaps
- Little bit of organisation here and there
- SceneManger object in scene now holds the spawner and unitManager (Any Prefab not spawned in scene needs to be a child of this)

+++ Emergence/Assets/Scripts/GridAStar.cs
@@ -125,7 +125,7 @@ public class GridAStar : MonoBehaviour
         {
             if (closedList.Contains(neighbourNode)/* || !neighbourNode.walkable*/) continue;
 
-            float tentativeGCost = currentNode.gCost + GetDistance(currentNode.cords, neighbourNode.cords);
+            int tentativeGCost = currentNode.gCost + GetDistance(currentNode.cords, neighbourNode.cords);
 
             if (tentativeGCost < neighbourNode.gCost)
             {
